The group enumeration seems to be a bit heavy when being used repeatedly for cell configuration. it actually gets into CoreData and use SQL query to get the images out from the Apple's database.
I tried to use the old idea which first load all assets into the array first, using the index range to get the partial of the array seems to be a lot more faster. There might be some overhead when we have 1000+ images, but it's still acceptable. We could think of putting an activity indicator somewhere to show user we are getting his pictures.